add pack stack subcommand
